Svelte 5 Migration 12:  Migrate user-facing pages + components to Svelte 5 + TS + all new features (#711)
* update some providers endpoints

* add `$utils` and `$icons` svelte definitions

* additional icons with TS

* update API + fetch types

* add `getCookie()` helper fn

* create a few new Svelte 5 + TS components

* update i18n

* update `/index`

* update `/acocunt` + `<AccInfo>`

* create whole new `<Input>` component

* create whole new `<Form>` component

* create internationalized date and time helpers

* `<InputDate>` calendar component + `<Form>` fixes

* Account info + edit tab migrated

* migrate `<PasswordPolicy>`

* fix casing for `UserAccountTypeResponse`

* CSS fixes for Options, Tabs, LangSelector

* create `<InputDateCombo>` to make `<InputDate>` nullable

* Webauthn rework WIP - Registration finished and working

* Webauthn rework WIP - Auth::Test working

* migrate `/device` UI

* rework of `webauthnAuth()` to make it more pluggable

* implement `Promise.race` for webauthn actions and show proper error messages in case of timeout

* UX improvements for `/device` page

* update i18n

* create `<Expandable>` component

* create `<LabeledValue>` component

* updates to OpenAPI docs

* rework `<Devices>` component

* rework `/logout` and create `post_dev_only_endpoints()` in the backend for improved DX

* rework `/authorize`

* rework `/error`

* rework `/providers/callback`

* move already migrated components into `lib_svelte5`

* impl `/dev/providers_callback` and fix some swagger docs

* fixed payload extraction for `post_dev_only_endpoints()` and added "register"

* updated i18n

* reworked `/users/register`

* reworked `/users/{id}/email_confirm`

* create the possibility to dynamically retrieve valid `tpl_password_reset` data during local dev

* create `<A>` component

* rework `/users/reset`

* create `<InputArea>`

* create `<Switch>`

* reworked `<AccWebId>`

* rework `<Tooltip>`

* various small CSS fixes

* rebuild UI

* clippy

+++ src/models/src/entity/auth_providers.rs
@@ -1093,7 +1093,7 @@ impl AuthProviderCallback {
                 header_origin,
                 user_id: user.id.clone(),
                 email: user.email,
-                exp: *WEBAUTHN_REQ_EXP,
+                exp: *WEBAUTHN_REQ_EXP as u64,
                 session,
             };
